Split the stages tests into their own class and condensed them - but maintaining the same code coverage.
